… docs/settings_page_optimization_branch_pr.md
… docs/settings_performance_optimization_fix.md
… memory_bank/memory_bank_active_context.md